pyproject.toml
[tool.poetry]
name = "lesterbot-v4"
version = "0.1.3"
description = ""
authors = ["Lukas Jaspaert <l.ester@gmx.de>"]
readme = "README.md"
packages = [{include = "lesterbot_v4"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py-interactions = "^5.10.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
pre-commit = "^3.5.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
